+++ settings.js
@@ -1,6 +1,8 @@
 export const registerSettings = function () {
 	// Register any custom module settings here
 	let modulename = "monks-scene-navigation";
+
+	const debouncedReload = foundry.utils.debounce(function () { window.location.reload(); }, 100);
 	
 	game.settings.register(modulename, "click-to-view", {
 		name: game.i18n.localize("MonksSceneNavigation.click-to-view.name"),
@@ -8,7 +10,7 @@ export const registerSettings = function () {
 		scope: "world",
 		config: true,
 		default: true,
-		type: Boolean,
+		type: Boolean
 	});
 	game.settings.register(modulename, "scene-indicator", {
 		name: game.i18n.localize("MonksSceneNavigation.scene-indicator.name"),
@@ -16,7 +18,16 @@ export const registerSettings = function () {
 		scope: "world",
 		config: true,
 		default: true,
+		type: Boolean
+	});
+	game.settings.register(modulename, "modify-scene-bar", {
+		name: game.i18n.localize("MonksSceneNavigation.modify-scene-bar.name"),
+		hint: game.i18n.localize("MonksSceneNavigation.modify-scene-bar.hint"),
+		scope: "world",
+		config: true,
+		default: true,
 		type: Boolean,
+		onChange: debouncedReload
 	});
 	game.settings.register(modulename, "folder-position", {
 		name: game.i18n.localize("MonksSceneNavigation.folder-position.name"),
@@ -36,7 +47,7 @@ export const registerSettings = function () {
 		scope: "world",
 		config: true,
 		default: false,
-		type: Boolean,
+		type: Boolean
 	});
 	game.settings.register(modulename, "display-background", {
 		name: game.i18n.localize("MonksSceneNavigation.display-background.name"),
@@ -45,8 +56,6 @@ export const registerSettings = function () {
 		config: true,
 		default: true,
 		type: Boolean,
-		onChange: () => {
-			location.reload();
-		}
+		onChange: debouncedReload
 	});
 };